requirements.txt
annotated-types==0.5.0
click==8.1.6
fastapi==0.100.1
FastAPI-SQLAlchemy==0.2.1
Flask==2.2.5
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.0.3
numpy==1.25.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
pandas==2.0.3
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.1
Werkzeug==2.3.6
